refactor: Update HTML template to use template literals
This commit refactors the HTML template in the code to use template literals. The previous version utilized JSX syntax, which has been replaced with template literals enclosed in backticks (``). This change makes it easier to embed dynamic values, such as ${projectName}, ${projectOverview}, ${hypothesis}, ${resultAndAnalysis}, and ${conclusion} within the HTML structure.